+++ example/error.ts
@@ -1,22 +1,7 @@
 import { Elysia, t } from '../src'
 
 new Elysia()
-	// .onError(({ code, error, set }) => {
-	// 	switch (code) {
-	// 		case 'NOT_FOUND':
-	// 			set.status = 404
-
-	// 			return 'Not Found :('
-
-	// 		case 'VALIDATION':
-	// 			set.status = 400
-
-	// 			return {
-	// 				fields: error.all
-	// 			}
-	// 	}
-	// })
-	.post('/a', async () => 'hi', {
+	.post('/', ({ body }) => body, {
 		body: t.Object({
 			username: t.String(),
 			password: t.String(),