+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 Improvement:
 - `AfterHandler` infer response type
 
+Bug fix:
+- [#1075](https://github.com/elysiajs/elysia/issues/1075)
+
 # 1.2.15 - 19 Feb 2025
 Bug fix:
 - [#1067](https://github.com/elysiajs/elysia/issues/1067) recompile async plugin once registered